requirements_dev.txt
-r requirements.txt

django-debug-toolbar
ipython
